TotalIn and TotalOut properties of IpfsStatsBw should be double
In order to support larger values

+++ src/Ipfs/Json/IpfsStatsBw.cs
@@ -3,8 +3,8 @@
 
     public class IpfsStatsBw
     {
-        public int TotalIn { get; set; }
-        public int TotalOut { get; set; }
+        public double TotalIn { get; set; }
+        public double TotalOut { get; set; }
         public double RateIn { get; set; }
         public double RateOut { get; set; }
     }
